+++ stress-tests/data-structures/RMQ.cpp
@@ -2,7 +2,7 @@
 
 #include "../../content/data-structures/RMQ.h"
 
-ll main() {
+int main() {
 	srand(2);
 	rep(N,0,100) {
 		vi v(N);
