+++ stress-tests/numerical/Tridiagonal.cpp
@@ -99,7 +99,7 @@ void rec(T& b, ll& a, F f) {
 	rep(i,0,mod) a = i, b = T(i), f();
 }
 
-ll main() {
+int main() {
 #ifdef BRUTEFORCE
 	const ll n = 3;
 	vector<vi> mat(n, vi(n)), mat2;
@@ -299,7 +299,7 @@ void rec(T& a, T& b, F f) {
 	}
 }
 
-ll main() {
+int main() {
 	ll count = 0;
 #ifdef BRUTEFORCE
 	const ll n = 3;
@@ -387,7 +387,7 @@ ll main() {
 
 }
 
-ll main() {
+int main() {
 	// finite_field::main();
 	real::main();
 }
